# docs/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "خالد سعد — عقد api/v1",
        "version": "1.0.0",
        "description": "مولَّدة من جدول المسارات — لا تُحرَّر يدويًّا.\n\nإصدارٌ واحد لا ثانٍ: يتطوّر api/v1 في مكانه، محروسًا ببوابة\nحدّ أدنى لإصدار التطبيق تُشحن قبل أي تغيير في العقد."
    },
    "servers": [
        {
            "url": "http://localhost/khaledsaad/public/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "kind",
                            "code",
                            "title",
                            "message",
                            "user_action"
                        ],
                        "properties": {
                            "kind": {
                                "type": "string",
                                "enum": [
                                    "ours",
                                    "theirs",
                                    "input"
                                ]
                            },
                            "code": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "user_action": {
                                "nullable": true,
                                "description": "يجب أن يكون null حين kind = ours.",
                                "type": "object",
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "route": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            },
                            "retry_after": {
                                "type": "integer",
                                "nullable": true
                            }
                        }
                    }
                }
            },
            "Envelope": {
                "type": "object",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object",
                        "properties": {
                            "locale": {
                                "type": "string"
                            },
                            "server_time": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "Success": {
                "description": "نجاح",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Envelope"
                        }
                    }
                }
            },
            "Error": {
                "description": "عطل مصنَّف",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/v1/admin/features/{feature}": {
            "put": {
                "operationId": "api.v1.admin.features.update",
                "summary": "api.v1.admin.features.update",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.admin.features.destroy",
                "summary": "api.v1.admin.features.destroy",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/admin/gateways/{gateway}": {
            "put": {
                "operationId": "api.v1.admin.gateways.update",
                "summary": "api.v1.admin.gateways.update",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.admin.gateways.destroy",
                "summary": "api.v1.admin.gateways.destroy",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
        "/api/v1/admin/gateways/{gateway}/default": {
            "patch": {
                "operationId": "api.v1.admin.gateways.default",
                "summary": "api.v1.admin.gateways.default",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/admin/packs/{pack}": {
            "put": {
                "operationId": "api.v1.admin.packs.update",
                "summary": "api.v1.admin.packs.update",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.admin.packs.destroy",
                "summary": "api.v1.admin.packs.destroy",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/admin/plans/{plan}": {
            "put": {
                "operationId": "api.v1.admin.plans.update",
                "summary": "api.v1.admin.plans.update",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.admin.plans.destroy",
                "summary": "api.v1.admin.plans.destroy",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/admin/tools/{tool}": {
            "get": {
                "operationId": "api.v1.admin.tools.show",
                "summary": "api.v1.admin.tools.show",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "put": {
                "operationId": "api.v1.admin.tools.update",
                "summary": "api.v1.admin.tools.update",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.admin.tools.destroy",
                "summary": "api.v1.admin.tools.destroy",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/competitors/{competitor}/confirm": {
            "post": {
                "operationId": "api.v1.competitors.confirm",
                "summary": "api.v1.competitors.confirm",
                "tags": [
                    "competitors"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/consultations/{consultation}": {
            "get": {
                "operationId": "api.v1.consultations.show",
                "summary": "api.v1.consultations.show",
                "tags": [
                    "consultations"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.consultations.destroy",
                "summary": "api.v1.consultations.destroy",
                "tags": [
                    "consultations"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/consultations/{consultation}/confirm": {
            "post": {
                "operationId": "api.v1.consultations.confirm",
                "summary": "api.v1.consultations.confirm",
                "tags": [
                    "consultations"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/consultations/{consultation}/evidence/{evidence}": {
            "delete": {
                "operationId": "api.v1.consultations.evidence.destroy",
                "summary": "api.v1.consultations.evidence.destroy",
                "tags": [
                    "consultations"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/devices": {
            "post": {
                "operationId": "api.v1.devices.store",
                "summary": "api.v1.devices.store",
                "tags": [
                    "devices"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "api.v1.devices.destroy",
                "summary": "api.v1.devices.destroy",
                "tags": [
                    "devices"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/public/runs/{run}/preflight": {
            "get": {
                "operationId": "api.v1.public.runs.preflight",
                "summary": "api.v1.public.runs.preflight",
                "tags": [
                    "public"
                ],
                "security": [],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/runs/{run}/files/{file}": {
            "delete": {
                "operationId": "api.v1.runs.files.destroy",
                "summary": "api.v1.runs.files.destroy",
                "tags": [
                    "runs"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
        "/api/v1/runs/{run}/insights": {
            "post": {
                "operationId": "api.v1.runs.insights",
                "summary": "api.v1.runs.insights",
                "tags": [
                    "runs"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/runs/{run}/preflight": {
            "get": {
                "operationId": "api.v1.runs.preflight",
                "summary": "api.v1.runs.preflight",
                "tags": [
                    "runs"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    }
}